+++ src/core/containers/AlleleFrequencyContainer.h
@@ -42,7 +42,7 @@ namespace transmission_nets::core::containers {
 
         void addLocus(std::shared_ptr<LocusImpl> locus);
 
-        double totalLoci() {
+        float totalLoci() {
             return alleleFrequencies_.size();
         }
