+++ src/core/datatypes/Matrix.h
@@ -13,16 +13,16 @@ namespace transmission_nets::core::datatypes {
     using SquareMatrix = Eigen::Matrix<T, DIM, DIM, Eigen::RowMajor>;
 
     template<int MAX_STATES>
-    using LogProbabilityTransitionMatrix = SquareMatrix<double, MAX_STATES>;
+    using LogProbabilityTransitionMatrix = SquareMatrix<float, MAX_STATES>;
 
     template<int MAX_STATES>
-    using TransitionMatrix = SquareMatrix<double, MAX_STATES>;
+    using TransitionMatrix = SquareMatrix<float, MAX_STATES>;
 
     template<int MAX_STATES>
-    using ProbabilityVector = Eigen::Matrix<double, MAX_STATES, 1>;
+    using ProbabilityVector = Eigen::Matrix<float, MAX_STATES, 1>;
 
-    using DynamicProbabilityVector = Eigen::Matrix<double, Eigen::Dynamic, 1>;
-    using DynamicArray             = Eigen::Array<double, Eigen::Dynamic, 1>;
+    using DynamicProbabilityVector = Eigen::Matrix<float, Eigen::Dynamic, 1>;
+    using DynamicArray             = Eigen::Array<float, Eigen::Dynamic, 1>;
 
     template<int MAX_COI>
     using COITransitionMatrix = LogProbabilityTransitionMatrix<MAX_COI>;
