+++ src/core/samplers/genetics/RandomAllelesBitSetSampler2.h
@@ -29,7 +29,7 @@ namespace transmission_nets::core::samplers::genetics {
 
         [[nodiscard]] unsigned int rejections() noexcept;
 
-        [[nodiscard]] double acceptanceRate() noexcept;
+        [[nodiscard]] float acceptanceRate() noexcept;
 
         [[nodiscard]] Likelihood logMetropolisHastingsAdjustment(const AllelesBitSetImpl& curr, const AllelesBitSetImpl& prop) noexcept;
 
@@ -100,8 +100,8 @@ namespace transmission_nets::core::samplers::genetics {
     }
 
     template<typename T, typename Engine, typename AllelesBitSetImpl, typename LocusImpl, typename ParentSetImpl>
-    double RandomAllelesBitSetSampler2<T, Engine, AllelesBitSetImpl, LocusImpl, ParentSetImpl>::acceptanceRate() noexcept {
-        return double(acceptances_) / (acceptances_ + rejections_);
+    float RandomAllelesBitSetSampler2<T, Engine, AllelesBitSetImpl, LocusImpl, ParentSetImpl>::acceptanceRate() noexcept {
+        return float(acceptances_) / (acceptances_ + rejections_);
     }
 
     template<typename T, typename Engine, typename AllelesBitSetImpl, typename LocusImpl, typename ParentSetImpl>
